Separate AlarmPlanViewModel.kt from AlarmEditViewModel.kt
