+++ app/src/main/java/com/quokkaman/android/app/common/data/AlarmDatePlan.kt
@@ -10,6 +10,4 @@ class AlarmDatePlan(private val date: Date) : AlarmPlan {
     }
 
     override fun getTitle(): String = SimpleDateFormat("MM월 dd일 (E)", Locale.KOREA).format(date.time)
-
-    override fun isDay(dayOfWeek: DayOfWeek): Boolean = false
 }
