+++ app/src/main/java/com/quokkaman/android/app/ui/alarm/edit/AlarmEditViewModel.kt
@@ -8,29 +8,7 @@ import java.util.*
 
 class AlarmEditViewModel : ViewModel() {
 
-    val planLiveData = MutableLiveData<AlarmPlan>().apply { value = AlarmDatePlan(Date()) }
-
     val alarmSoundSettingLiveData: LiveData<AlarmSetting> = MutableLiveData()
     val alarmVibrateSettingLiveData: LiveData<AlarmSetting> = MutableLiveData()
     val alarmRepeatSettingLiveData: LiveData<AlarmSetting> = MutableLiveData()
-
-    fun clickDate() {
-        val alarmDatePlan = AlarmDatePlan(Date())
-        planLiveData.postValue(alarmDatePlan)
-    }
-
-    fun clickDay(dayOfWeek: DayOfWeek) {
-        val prevPlan = planLiveData.value
-        val alarmPlan = if (prevPlan !is AlarmDayPlan) {
-            AlarmDayPlan(EnumSet.of(dayOfWeek), false)
-        } else {
-            prevPlan.toggle(dayOfWeek)
-            if (prevPlan.isEmpty()) {
-                AlarmDatePlan(Date())
-            } else {
-                prevPlan
-            }
-        }
-        planLiveData.postValue(alarmPlan)
-    }
 }
